func-sinistro/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.18.0
python-dotenv==1.0.1
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-ai-formrecognizer==3.3.2
azure-ai-vision-imageanalysis==1.0.0b2
extract_msg==0.48.0